Update and rename Akhilesh_Electric Vehicle_EDA.ipynb to Project 2.1 Electric Vehicle.ipynb
I work for an electric vehicle (EV) start-up. The start-up is evaluating which vehicle/customer segment to target for developing its EVs. Our task is to analyze the electric vehicle market in India using segmentation analysis. Based on this analysis, we need to formulate a feasible strategy to enter the market by targeting the segments most likely to adopt electric vehicles.